Refine the TEE proof submission endpoint
