+++ core/node/proof_data_handler/src/tee_request_processor.rs
@@ -1,20 +1,16 @@
 use std::sync::Arc;
 
-use axum::{
-    extract::Path,
-    http::StatusCode,
-    response::{IntoResponse, Response},
-    Json,
-};
+use crate::errors::RequestProcessorError;
+use axum::{extract::Path, Json};
 use zksync_config::configs::ProofDataHandlerConfig;
-use zksync_dal::{ConnectionPool, Core, CoreDal, SqlxError};
-use zksync_object_store::{ObjectStore, ObjectStoreError};
+use zksync_dal::{ConnectionPool, Core, CoreDal};
+use zksync_object_store::ObjectStore;
 use zksync_prover_interface::api::{
-    GenericProofGenerationDataResponse, SubmitProofRequest, SubmitProofResponse,
+    GenericProofGenerationDataResponse, SubmitProofResponse, SubmitTeeProofRequest,
     TeeProofGenerationDataRequest,
 };
 use zksync_tee_verifier::TeeVerifierInput;
-use zksync_types::{commitment::serialize_commitments, web3::keccak256, L1BatchNumber, H256};
+use zksync_types::L1BatchNumber;
 
 pub type TeeProofGenerationDataResponse = GenericProofGenerationDataResponse<TeeVerifierInput>;
 
@@ -23,38 +19,6 @@ pub(crate) struct TeeRequestProcessor {
     blob_store: Arc<dyn ObjectStore>,
     pool: ConnectionPool<Core>,
     config: ProofDataHandlerConfig,
-}
-
-pub(crate) enum TeeRequestProcessorError {
-    ObjectStore(ObjectStoreError),
-    Sqlx(SqlxError),
-}
-
-impl IntoResponse for TeeRequestProcessorError {
-    fn into_response(self) -> Response {
-        let (status_code, message) = match self {
-            TeeRequestProcessorError::ObjectStore(err) => {
-                tracing::error!("GCS error: {:?}", err);
-                (
-                    StatusCode::BAD_GATEWAY,
-                    "Failed fetching/saving from GCS".to_owned(),
-                )
-            }
-            TeeRequestProcessorError::Sqlx(err) => {
-                tracing::error!("Sqlx error: {:?}", err);
-                match err {
-                    SqlxError::RowNotFound => {
-                        (StatusCode::NOT_FOUND, "Non existing L1 batch".to_owned())
-                    }
-                    _ => (
-                        StatusCode::BAD_GATEWAY,
-                        "Failed fetching/saving from db".to_owned(),
-                    ),
-                }
-            }
-        };
-        (status_code, message).into_response()
-    }
 }
 
 impl TeeRequestProcessor {
@@ -73,7 +37,7 @@ impl TeeRequestProcessor {
     pub(crate) async fn get_proof_generation_data(
         &self,
         request: Json<TeeProofGenerationDataRequest>,
-    ) -> Result<Json<TeeProofGenerationDataResponse>, TeeRequestProcessorError> {
+    ) -> Result<Json<TeeProofGenerationDataResponse>, RequestProcessorError> {
         tracing::info!("Received request for proof generation data: {:?}", request);
 
         // TODO: Replace this line with an appropriate method to get the next batch number to be proven.
@@ -90,7 +54,7 @@ impl TeeRequestProcessor {
             .blob_store
             .get(l1_batch_number)
             .await
-            .map_err(TeeRequestProcessorError::ObjectStore)?;
+            .map_err(RequestProcessorError::ObjectStore)?;
 
         Ok(Json(TeeProofGenerationDataResponse::Success(Some(
             Box::new(tee_verifier_input),
@@ -100,104 +64,25 @@ impl TeeRequestProcessor {
     pub(crate) async fn submit_proof(
         &self,
         Path(l1_batch_number): Path<u32>,
-        Json(payload): Json<SubmitProofRequest>,
-    ) -> Result<Json<SubmitProofResponse>, TeeRequestProcessorError> {
+        Json(payload): Json<SubmitTeeProofRequest>,
+    ) -> Result<Json<SubmitProofResponse>, RequestProcessorError> {
         tracing::info!("Received proof for block number: {:?}", l1_batch_number);
+
         let l1_batch_number = L1BatchNumber(l1_batch_number);
+        let mut connection = self.pool.connection().await.unwrap();
+        let mut dal = connection.proof_generation_dal();
+
+        // TODO: Replace the lines below with code that saves the proof generation result back to the database.
         match payload {
-            SubmitProofRequest::Proof(proof) => {
-                let blob_url = self
-                    .blob_store
-                    .put(l1_batch_number, &*proof)
-                    .await
-                    .map_err(TeeRequestProcessorError::ObjectStore)?;
-
-                let system_logs_hash_from_prover =
-                    H256::from_slice(&proof.aggregation_result_coords[0]);
-                let state_diff_hash_from_prover =
-                    H256::from_slice(&proof.aggregation_result_coords[1]);
-                let bootloader_heap_initial_content_from_prover =
-                    H256::from_slice(&proof.aggregation_result_coords[2]);
-                let events_queue_state_from_prover =
-                    H256::from_slice(&proof.aggregation_result_coords[3]);
-
-                let mut storage = self.pool.connection().await.unwrap();
-
-                let l1_batch = storage
-                    .blocks_dal()
-                    .get_l1_batch_metadata(l1_batch_number)
-                    .await
-                    .unwrap()
-                    .expect("Proved block without metadata");
-
-                let is_pre_boojum = l1_batch
-                    .header
-                    .protocol_version
-                    .map(|v| v.is_pre_boojum())
-                    .unwrap_or(true);
-                if !is_pre_boojum {
-                    let events_queue_state = l1_batch
-                        .metadata
-                        .events_queue_commitment
-                        .expect("No events_queue_commitment");
-                    let bootloader_heap_initial_content = l1_batch
-                        .metadata
-                        .bootloader_initial_content_commitment
-                        .expect("No bootloader_initial_content_commitment");
-
-                    if events_queue_state != events_queue_state_from_prover
-                        || bootloader_heap_initial_content
-                            != bootloader_heap_initial_content_from_prover
-                    {
-                        let server_values = format!("events_queue_state = {events_queue_state}, bootloader_heap_initial_content = {bootloader_heap_initial_content}");
-                        let prover_values = format!("events_queue_state = {events_queue_state_from_prover}, bootloader_heap_initial_content = {bootloader_heap_initial_content_from_prover}");
-                        panic!(
-                            "Auxilary output doesn't match, server values: {} prover values: {}",
-                            server_values, prover_values
-                        );
-                    }
-                }
-
-                let system_logs = serialize_commitments(&l1_batch.header.system_logs);
-                let system_logs_hash = H256(keccak256(&system_logs));
-
-                if !is_pre_boojum {
-                    let state_diff_hash = l1_batch
-                        .header
-                        .system_logs
-                        .into_iter()
-                        .find(|elem| elem.0.key == H256::from_low_u64_be(2))
-                        .expect("No state diff hash key")
-                        .0
-                        .value;
-
-                    if state_diff_hash != state_diff_hash_from_prover
-                        || system_logs_hash != system_logs_hash_from_prover
-                    {
-                        let server_values = format!("system_logs_hash = {system_logs_hash}, state_diff_hash = {state_diff_hash}");
-                        let prover_values = format!("system_logs_hash = {system_logs_hash_from_prover}, state_diff_hash = {state_diff_hash_from_prover}");
-                        panic!(
-                            "Auxilary output doesn't match, server values: {} prover values: {}",
-                            server_values, prover_values
-                        );
-                    }
-                }
-                storage
-                    .proof_generation_dal()
-                    .save_proof_artifacts_metadata(l1_batch_number, &blob_url)
-                    .await
-                    .map_err(TeeRequestProcessorError::Sqlx)?;
+            SubmitTeeProofRequest::Proof(_proof) => {
+                // dal.save_proof_artifacts_metadata(l1_batch_number, &blob_url)
+                //     .await
+                //     .map_err(RequestProcessorError::Sqlx)?;
             }
-            SubmitProofRequest::TeeProof(_proof) => { /* TBD */ }
-            SubmitProofRequest::SkippedProofGeneration => {
-                self.pool
-                    .connection()
+            SubmitTeeProofRequest::SkippedProofGeneration => {
+                dal.mark_proof_generation_job_as_skipped(l1_batch_number)
                     .await
-                    .unwrap()
-                    .proof_generation_dal()
-                    .mark_proof_generation_job_as_skipped(l1_batch_number)
-                    .await
-                    .map_err(TeeRequestProcessorError::Sqlx)?;
+                    .map_err(RequestProcessorError::Sqlx)?;
             }
         }
